Updating to the new blazor template for blazor interactive support.

+++ src/AlpacaHerderOld/Shared/Quote.cs
@@ -4,11 +4,11 @@ namespace AlpacaHerder.Shared {
     public record Quote {
         public string Symbol { get; init; }
 
-        public ulong BidSize { get; init; }
+        public decimal BidSize { get; init; }
 
         public decimal BidPrice { get; init; }
 
-        public ulong AskSize { get; init; }
+        public decimal AskSize { get; init; }
 
         public decimal AskPrice { get; init; }
